- Update Zod schemas to enforce integer amounts for `initialBalance` (preventing float precision issues).
- Add maximum length constraints to `notes`, `RuleCondition`, and `RuleAction` values to mitigate DoS risks.
- Ensure numeric rule values are finite.
- Implement DNS rebinding protection in `StreamableHTTPHandler` and `index.ts`.
- Restrict `allowedHosts` to `localhost` when bearer authentication is disabled to prevent unauthorized external access.
- Replace `Object.prototype.hasOwnProperty.call` with `Object.hasOwn` for safer property checks.
- Update `pnpm-lock.yaml` to reflect dependency changes.

+++ mcp-server/src/tools/crud-factory-config.ts
@@ -128,7 +128,7 @@ const CreateAccountSchema = z.object({
     }),
   }),
   offbudget: z.boolean().optional(),
-  initialBalance: z.number().optional(), // In cents
+  initialBalance: z.number().int().optional(), // In cents
 });
 
 const UpdateAccountSchema = z.object({
